+++ src/components/ModelProviderIcon/index.tsx
@@ -1,5 +1,6 @@
 import {
   Ai360,
+  AiMass,
   Anthropic,
   Azure,
   Baichuan,
@@ -118,6 +119,10 @@ const ModelProviderIcon = memo<ModelProviderIconProps>(({ provider }) => {
 
     case ModelProvider.Baichuan: {
       return <Baichuan size={20} />;
+    }
+
+    case ModelProvider.Taichu: {
+      return <AiMass size={20} />;
     }
 
     case ModelProvider.Zhinao: {