+++ src/const/settings/llm.ts
@@ -14,6 +14,7 @@ import {
   PerplexityProviderCard,
   QwenProviderCard,
   StepfunProviderCard,
+  TaichuProviderCard,
   TogetherAIProviderCard,
   ZeroOneProviderCard,
   ZhinaoProviderCard,
@@ -87,6 +88,10 @@ export const DEFAULT_LLM_CONFIG: UserModelProviderConfig = {
   stepfun: {
     enabled: false,
     enabledModels: filterEnabledModels(StepfunProviderCard),
+  },
+  taichu: {
+    enabled: false,
+    enabledModels: filterEnabledModels(TaichuProviderCard),
   },
   togetherai: {
     enabled: false,