+++ src/features/Conversation/Error/APIKeyForm/ProviderAvatar.tsx
@@ -1,5 +1,6 @@
 import {
   Ai360,
+  AiMass,
   Anthropic,
   Baichuan,
   DeepSeek,
@@ -80,6 +81,10 @@ const ProviderAvatar = memo<ProviderAvatarProps>(({ provider }) => {
 
     case ModelProvider.Stepfun: {
       return <Stepfun color={Stepfun.colorPrimary} size={56} />;
+    }
+
+    case ModelProvider.Taichu: {
+      return <AiMass.Color size={56} />;
     }
 
     case ModelProvider.TogetherAI: {