+++ src/layout/AuthProvider/Clerk/index.tsx
@@ -4,10 +4,13 @@ import { ClerkProvider } from '@clerk/nextjs';
 import { PropsWithChildren, memo, useEffect, useMemo, useState, useTransition } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import { featureFlagsSelectors, useServerConfigStore } from '@/store/serverConfig';
+
 import UserUpdater from './UserUpdater';
 import { useAppearance } from './useAppearance';
 
 const Clerk = memo(({ children }: PropsWithChildren) => {
+  const { enableClerkSignUp } = useServerConfigStore(featureFlagsSelectors);
   const appearance = useAppearance();
   const {
     i18n: { language, getResourceBundle },
@@ -18,17 +21,32 @@ const Clerk = memo(({ children }: PropsWithChildren) => {
   // When useAppearance returns different result during SSR vs. client-side (when theme mode is auto), the appearance is not applied
   // It's because Clerk internally re-applies SSR props after transition which overrides client-side props, see https://github.com/clerk/javascript/blob/main/packages/nextjs/src/app-router/client/ClerkProvider.tsx
   // This re-renders the provider after transition to make sure client-side props are always applied
-  const [, setCount] = useState(0);
+  const [count, setCount] = useState(0);
   const [isPending, startTransition] = useTransition();
   useEffect(() => {
-    if (isPending) return;
+    if (count || isPending) return;
     startTransition(() => {
       setCount((count) => count + 1);
     });
-  }, [isPending, startTransition]);
+  }, [count, setCount, isPending, startTransition]);
+
+  const updatedAppearance = useMemo(
+    () => ({
+      ...appearance,
+      elements: {
+        ...appearance.elements,
+        ...(!enableClerkSignUp ? { footerAction: { display: 'none' } } : {}),
+      },
+    }),
+    [appearance, enableClerkSignUp],
+  );
 
   return (
-    <ClerkProvider appearance={appearance} localization={localization}>
+    <ClerkProvider
+      appearance={updatedAppearance}
+      localization={localization}
+      signUpUrl={!enableClerkSignUp ? '/login' : '/signup'} // Redirect sign-up to sign-in if disabled
+    >
       {children}
       <UserUpdater />
     </ClerkProvider>